Revert to old bounce filters, keep cushions/maxbounce
With the current setup, only gravity-affected entities are eligible to be affected by trampolines, which is intuitive behavior, whereas rockets/nails, though fun for sure, are more questionable (why does the rocket bounce instead of just mangling the trampoline/destroying it?). The exclusion of nails only adds to the complexity of what is an eligible bouncee.

Cushion-only and min speed settings kept, as well as maxbounce.