Fix a bug in pd3.0 key doors
The custom failure message set by the mapper was ignored in case of a key door (actually instantly replaced by the default message).

The door_touch behavior was:
1. Print the custom failure message with a "talk" sound no matter if the door requires a key or not.
2. If the door requires a key, then use "" as custom failure message.
3. If the custom failure message is "" print the default failure message (with a "key missing" sound).
So at the end what the player saw was the default failure message.

New door_touch behavior:
1. Print the custom failure message with the "talk" sound only for doors not requiring a key.
2. If the door requires a key, then use the mapper's custom failure message (as it was not already printed at the previous step).
3. Print it with the "key missing" sound.
So now what the player see is the expected failure message.