Force state for func_togglevisiblewall
To avoid messy situation where you lost track of how many times the thing was toggled and whether it's currently ON or OFF (and to avoid super complex setups involving tons of side entities to figure out), this PR allows to force the state to either ON or OFF by means of target_setstate (which already does the same for other entitires).

Also incorporated latest edits made in progs_dump to the target_setstate code to support lights.

Also added pain_target support to trigger_look in the same spirit as it's used in other entities: targets fired when the trigger's base condition is NOT met (i.e. when the player is NOT looking at the trigger's target).