"Inventory empty" caption
A caption was added to tell the player their puzzle items inventory is empty.
Previously, only the inventory plaque was displayed and nothing more. This could lead the player to think that something was missing/f***ed up.
Or even, they could have grabbed and used puzzle items even realizing an inventory exists. Their first encounter with it would then be extremely… puzzling! (Carcosa was especially prone to that)